+++ admin_board_view/templates/products.html
@@ -3,12 +3,11 @@
 {% block body %}
 <section class="d-flex">
   <h1 class="flex-grow-1">Products</h1>
-  <button class="btn btn-success" id="addProduct" data-bs-toggle="modal" data-bs-target="#addProductModal">Add product</button>
+  <a href="/products?id=0"><button class="btn btn-success">Add product</button></a>
 </section> 
 <table class="table table-striped table-hover text-center align-middle">
   <thead>
     <tr>
-      <th>Select</th>
       <th>Image</th>
       <th>Name</th>
       <th>Price</th>
@@ -19,13 +18,12 @@
   <tbody>
     {% for product in products %}
       <tr id="{{ product.id }}">
-        <td><input type="checkbox"/></td>
         <td class="product-image {% if not product.enabled %}disabled-product{% endif %}">{{ product.image_view }}</td>
         <td>{{ product.name }}</td>
         <td>{{ product.euro }}</td>
         <td>{{ product.category }}</td>
         <td>
-          <button class="btn btn-primary edit-product">Edit</button>
+          <a href="/products?id={{ product.id }}"><button class="btn btn-primary edit-product">Edit</button></a>
           <button class="btn btn-warning toggle-product" name="{{ product.id }}">{% if product.enabled %}De{% else %}Re{% endif %}activate</button>
           <button class="btn btn-danger delete-product" name="{{ product.id }}">Delete</button>
         </td>
@@ -33,26 +31,31 @@
     {% endfor %}
   </tbody>
 </table>
-<div class="modal fade" id="addProductModal" tabindex="-1">
+<div class="modal fade" id="productModal" tabindex="-1">
   <div class="modal-dialog">
     <div class="modal-content">
       <form method="post" enctype="multipart/form-data">
       <div class="modal-header">
-        <h5 class="modal-title">Add new product</h5>
+        <h5 class="modal-title">Add or edit product</h5>
         <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
       </div>
       <div class="modal-body">
         {% csrf_token %}
-        {{ image_upload.as_p }}
+        {{ product_form.as_p }}
+        {% if current_product %}
+          <label class="form-label">Current image</label>
+          {{ current_product.image_view }}
+        {% endif %}
       </div>
       <div class="modal-footer">
         <div class="input-group mb-3 btn-group">
           <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-          <button type="submit" class="btn btn-success">Add product</button>
+          <button type="submit" class="btn btn-success">Save product</button>
         </div>
       </div>
       </form>
     </div>
   </div>
 </div>
+<script type="text/javascript" src="{% static 'AdminBoardView/products.js' %}" defer></script>
 {% endblock %}
